+++ FunKiiUNETThingy/frmSettings.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.ComponentModel;
 using System.Data;
 using System.Drawing;
+using System.IO;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
@@ -12,8 +13,8 @@ namespace FunKiiUNETThingy
 {
     public partial class frmSettings : Form
     {
-        private Config config;
-        private string cfgFileName;
+        public Config config;
+        //private string cfgFileName;
 
         public frmSettings(Config _config, string fileName)
         {
@@ -21,7 +22,7 @@ namespace FunKiiUNETThingy
             this.Icon = FunKiiUNETThingy.Properties.Resources.FunKiiUNETThingy;
 
             config = _config;
-            cfgFileName = fileName;
+            //cfgFileName = fileName;
         }
 
         private void frmSettings_Load(object sender, EventArgs e)
@@ -30,6 +31,9 @@ namespace FunKiiUNETThingy
             chkAutoLoadData.Checked = config.appAutoLoadData;
             chkFileDownloadSkip.Checked = config.appDlIgnoreExistingContentFiles;
             chkGroupDownloads.Checked = config.appDlGroupDlsIntoSubfolders;
+            txtSaveDir.Text = config.saveDir;
+
+            chkFilesize1024.Checked = config.appFilesize1024;
         }
 
         private void btnSaveSettings_Click(object sender, EventArgs e)
@@ -38,23 +42,41 @@ namespace FunKiiUNETThingy
             config.appAutoLoadData = chkAutoLoadData.Checked;
             config.appDlIgnoreExistingContentFiles = chkFileDownloadSkip.Checked;
             config.appDlGroupDlsIntoSubfolders = chkGroupDownloads.Checked;
+            config.appFilesize1024 = chkFilesize1024.Checked;
 
-            try
+            if (Directory.Exists(txtSaveDir.Text))
+                config.saveDir = txtSaveDir.Text;
+            else
             {
-                config.SaveToFile(cfgFileName);
-                lblSaveResult.ForeColor = Color.DarkGreen;
-                lblSaveResult.Text = cfgFileName + " saved successfully!";
-            }
-            catch (Exception ex)
-            {
+                txtSaveDir.BackColor = Color.LightSalmon;
                 lblSaveResult.ForeColor = Color.Red;
-                lblSaveResult.Text = String.Format("ERROR! Failed to save {0}!" + Environment.NewLine + "{1}", cfgFileName, ex.Message);
-            }    
+                lblSaveResult.Text = "Please enter/select a valid Downloads Directory!";
+                return;
+            }
+
+            this.DialogResult = DialogResult.OK;
+            this.Close();
         }
 
         private void btnCancel_Click(object sender, EventArgs e)
         {
+            this.DialogResult = DialogResult.Cancel;
             this.Close();
+        }
+
+        private void btnDirSelect_Click(object sender, EventArgs e)
+        {
+            using (FolderBrowserDialog fbd = new FolderBrowserDialog())
+            {
+                DialogResult result = fbd.ShowDialog();
+
+                if (result == DialogResult.OK)
+                {
+                    txtSaveDir.Text = fbd.SelectedPath;
+                    txtSaveDir.BackColor = Color.White;
+                    lblSaveResult.Text = "";
+                }
+            }
         }
     }
 }